requirements.txt
argcomplete==1.11.1
beautifulsoup4==4.6.0
boto==2.49.0
boto3==1.12.6
botocore==1.15.6
certifi==2019.11.28
cfn-flip==1.2.2
chardet==3.0.4
Click==7.0
Django==2.1
django-modelcluster==5.0.1
django-s3-sqlite==0.0.3
django-s3-storage==0.13.0
django-storages==1.9.1
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
docutils==0.15.2
draftjs-exporter==2.1.7
durationpy==0.5
future==0.18.2
hjson==3.0.1
html5lib==1.0.1
idna==2.9
importlib-metadata==1.5.0
jmespath==0.9.5
kappa==0.6.0
Pillow==6.2.2
pip-tools==4.5.0
placebo==0.9.0
python-dateutil==2.6.1
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
toml==0.10.0
tqdm==4.43.0
troposphere==2.6.0
Unidecode==1.1.1
urllib3==1.25.8
wagtail==2.7.1
webencodings==0.5.1
Werkzeug==1.0.0
Willow==1.3
wsgi-request-logger==0.4.6
zappa==0.50.0
zappa-django-utils==0.4.0
zipp==3.0.0


